@@ -1,7 +1,7 @@
 {
-  "name": "turtlecoin-utils",
-  "version": "2.0.0",
-  "description": "A Node.js collection of utilities useful for TurtleCoin developers",
+  "name": "wrkzcoin-utils",
+  "version": "3.0.3",
+  "description": "A Typescript/CommonJS collection of utilities useful for WrkzCoin developers",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
@@ -13,53 +13,70 @@
   },
   "scripts": {
     "build": "./node_modules/.bin/tsc",
-    "docs": "./node_modules/.bin/typedoc && touch docs/.nojekyll",
+    "docs": "./node_modules/.bin/typedoc && touch docs/.nojekyll && git restore docs/CNAME",
     "prepare": "npm run build",
     "prepublishOnly": "npm test && npm run docs",
-    "style": "./node_modules/.bin/tslint -c tslint.json --project .",
-    "fix-style": "./node_modules/.bin/tslint -c tslint.json --fix --project .",
+    "style": "./node_modules/.bin/eslint src/**/*.ts src/*.ts test/*.ts",
+    "fix-style": "./node_modules/.bin/eslint --fix src/**/*.ts src/*.ts test/*.ts",
     "test": "npm run style && npm run build && npm run test-node && npm run test-js",
-    "mocha": "./node_modules/.bin/mocha",
-    "test-node": "npm run mocha",
-    "test-js": "env FORCE_JS=1 npm run mocha"
+    "test-node": "./node_modules/.bin/mocha --exit --timeout 30000 --require ts-node/register test/test.ts",
+    "test-js": "env FORCE_JS=1 ./node_modules/.bin/mocha --exit --timeout 30000 --require ts-node/register test/test.ts"
   },
   "keywords": [
-    "TurtleCoin",
+    "WrkzCoin",
     "Utilities"
   ],
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/turtlecoin/turtlecoin-utils.git"
+    "url": "git+https://github.com/wrkzcoin/wrkzcoin-utils.git"
   },
-  "homepage": "https://utils.turtlecoin.dev",
+  "homepage": "https://wrkzcoin-utils.wrkz.work",
   "bugs": {
-    "url": "https://github.com/turtlecoin/turtlecoin-utils/issues"
+    "url": "https://github.com/wrkzcoin/wrkzcoin-utils/issues"
   },
   "author": "Brandon Lehmann <brandonlehmann@gmail.com>",
   "contributors": [
     "zpalmtree <22151537+zpalmtree@users.noreply.github.com>"
   ],
   "license": "GPL-3.0",
+  "optionalDependencies": {
+    "@ledgerhq/hw-transport-node-hid": "^5.22.0"
+  },
   "dependencies": {
+    "@types/node-fetch": "^2.5.7",
+    "@types/ledgerhq__hw-transport": "^4.21.2",
+    "@ledgerhq/hw-transport": "^5.22.0",
+    "abort-controller": "^3.0.0",
     "aes-js": "^3.1.2",
     "big-integer": "^1.6.48",
-    "bytestream-helper": "^0.0.7",
+    "bytestream-helper": "0.0.10",
     "get-random-values": "^1.2.0",
+    "node-fetch": "^2.6.0",
     "numeral": "^2.0.6",
-    "secure-random-string": "^1.1.3",
-    "turtlecoin-base58": "^0.0.5",
-    "turtlecoin-crypto": "^4.0.6",
-    "turtlecoin-mnemonics": "^0.0.5"
+    "semver": "^7.3.2",
+    "turtlecoin-base58": "0.0.8",
+    "wrkzcoin-crypto": "https://github.com/wrkzcoin/wrkzcoin-crypto#cn_trtl_tx_pow",
+    "turtlecoin-mnemonics": "0.0.9"
   },
   "devDependencies": {
     "@types/aes-js": "^3.1.1",
-    "@types/node": "^13.13.0",
-    "@types/numeral": "0.0.26",
-    "mocha": "^7.1.1",
-    "tslint": "^6.1.1",
-    "typedoc": "^0.17.4",
-    "typescript": "^3.8.3",
-    "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11"
+    "@types/ledgerhq__hw-transport": "^4.21.2",
+    "@types/ledgerhq__hw-transport-node-hid": "^4.22.1",
+    "@types/mocha": "^8.0.3",
+    "@types/node": "^14.6.0",
+    "@types/numeral": "0.0.28",
+    "@types/semver": "^7.3.3",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
+    "eslint": "^7.7.0",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
+    "mocha": "^8.1.1",
+    "ts-node": "^8.10.2",
+    "typedoc": "^0.18.0",
+    "typescript": "^3.9.7"
   }
 }
