@@ -2,15 +2,9 @@
 //
 // Please see the included LICENSE file for more information.
 
-import {Common} from '../Common';
-import {TurtleCoinCrypto} from '../Types';
-import {Mnemonics} from 'turtlecoin-mnemonics';
-
-/** @ignore */
-const Config = require('../../config.json');
-
-/** @ignore */
-const SecureRandomString = require('secure-random-string');
+import { Config } from '../Config';
+import { TurtleCoinCrypto } from '../Types';
+import { randomBytes } from 'crypto';
 
 export namespace ED25519 {
     /**
@@ -18,8 +12,8 @@ export namespace ED25519 {
      * for generating new key pairs including deterministic methods.
      */
     export class KeyPair {
-        private m_privateKey?: string;
-        private m_publicKey?: string;
+        protected m_privateKey?: string;
+        protected m_publicKey?: string;
 
         /**
          * Constructs a new KeyPair object
@@ -27,40 +21,59 @@ export namespace ED25519 {
          * @param privateKey
          * @param entropy
          * @param iterations
+         * @param createEmpty
          */
-        constructor(publicKey?: string, privateKey?: string, entropy?: string, iterations?: number) {
-            /* If no entropy was supplied, we'll go find our own */
-            entropy = entropy || SecureRandomString({length: 256});
+        public static async from (
+            publicKey?: string,
+            privateKey?: string,
+            entropy?: string,
+            iterations?: number,
+            createEmpty = false
+        ): Promise<KeyPair> {
+            const pair = new KeyPair();
 
-            if (publicKey && TurtleCoinCrypto.checkKey(publicKey)) {
-                this.m_publicKey = publicKey;
+            if (createEmpty) {
+                return pair;
             }
 
-            if (privateKey && TurtleCoinCrypto.checkScalar(privateKey)) {
-                this.m_privateKey = privateKey;
+            /* If no entropy was supplied, we'll go find our own */
+            entropy = entropy || rand(256);
+
+            if (publicKey && await TurtleCoinCrypto.checkKey(publicKey)) {
+                pair.m_publicKey = publicKey;
+            }
+
+            if (privateKey && await TurtleCoinCrypto.checkScalar(privateKey)) {
+                pair.m_privateKey = privateKey;
             }
 
             if (!publicKey && !privateKey) {
-                this.privateKey = simpleKdf(
-                    entropy + rand32(), iterations || Config.keccakIterations);
+                const temp = await simpleKdf(
+                    entropy + rand(32), iterations || Config.keccakIterations);
+                await pair.setPrivateKey(temp);
             }
 
-            if (this.m_privateKey && !this.m_publicKey) {
+            if (pair.m_privateKey && !pair.m_publicKey) {
                 /* If we supplied a private key but no public key, and we said 1 iteration
-                we are probably looking to generate the deterministic view key for the
-                specified private spend key */
+                    we are probably looking to generate the deterministic view key for the
+                    specified private spend key */
                 if (iterations && iterations === 1) {
-                    this.privateKey = TurtleCoinCrypto.cn_fast_hash(this.m_privateKey);
+                    const temp = await TurtleCoinCrypto.cn_fast_hash(
+                        pair.m_privateKey);
+
+                    await pair.setPrivateKey(temp);
                 }
 
-                this.m_publicKey = TurtleCoinCrypto.secretKeyToPublicKey(this.m_privateKey);
+                pair.m_publicKey = await TurtleCoinCrypto.secretKeyToPublicKey(pair.m_privateKey);
             }
+
+            return pair;
         }
 
         /**
          * Returns the private key
          */
-        public get privateKey(): string {
+        public get privateKey (): string {
             return (this.m_privateKey) ? this.m_privateKey : '';
         }
 
@@ -68,15 +81,19 @@ export namespace ED25519 {
          * Sets the private key or reduces the value to a private key
          * @param key
          */
-        public set privateKey(key: string) {
-            this.m_privateKey = (TurtleCoinCrypto.checkScalar(key)) ?
-                key : TurtleCoinCrypto.scReduce32(key);
+        public async setPrivateKey (key: string): Promise<void> {
+            try {
+                this.m_privateKey = (await TurtleCoinCrypto.checkScalar(key))
+                    ? key : await TurtleCoinCrypto.scReduce32(key);
+            } catch (e) {
+                this.m_publicKey = key;
+            }
         }
 
         /**
          * Returns the public key
          */
-        public get publicKey(): string {
+        public get publicKey (): string {
             return (this.m_publicKey) ? this.m_publicKey : '';
         }
 
@@ -84,8 +101,20 @@ export namespace ED25519 {
          * Sets the public key
          * @param key
          */
-        public set publicKey(key: string) {
-            if (TurtleCoinCrypto.checkKey(key)) {
+        public async setPublicKey (key: string): Promise<void> {
+            let isPubKey = false;
+
+            // Try to verify that it is a public key via the library
+            try {
+                isPubKey = await TurtleCoinCrypto.checkKey(key);
+            } catch (e) {
+                // If the library could not process this, then set the key anyway
+                this.m_publicKey = key;
+
+                return;
+            }
+
+            if (isPubKey) {
                 this.m_publicKey = key;
             } else {
                 throw new Error('Not a public key');
@@ -95,12 +124,12 @@ export namespace ED25519 {
         /**
          * Returns if the public key belongs to the private key
          */
-        public get isPaired(): boolean {
+        public async isPaired (): Promise<boolean> {
             if (this.publicKey.length === 0 || this.privateKey.length === 0) {
                 return false;
             }
 
-            return (TurtleCoinCrypto.secretKeyToPublicKey(this.privateKey) === this.publicKey);
+            return (await TurtleCoinCrypto.secretKeyToPublicKey(this.privateKey) === this.publicKey);
         }
     }
 
@@ -108,33 +137,38 @@ export namespace ED25519 {
      * Represents a set of ED25519 key pairs (view and spend) used by TurtleCoin wallets
      */
     export class Keys {
-        private m_spendKeys: KeyPair = new KeyPair();
-        private m_viewKeys: KeyPair = new KeyPair();
+        protected m_spendKeys: KeyPair = new KeyPair();
+        protected m_viewKeys: KeyPair = new KeyPair();
 
         /**
          * Creates a new instance of a set of Keys
          * @param spendKeys the spend key pair
          * @param viewKeys the view key pair
          */
-        constructor(spendKeys?: KeyPair, viewKeys?: KeyPair) {
+        public static async from (spendKeys?: KeyPair, viewKeys?: KeyPair): Promise<Keys> {
+            const keys = new Keys();
+
             if (spendKeys) {
-                this.m_spendKeys = spendKeys;
+                keys.m_spendKeys = spendKeys;
             }
 
             if (viewKeys) {
-                this.m_viewKeys = viewKeys;
+                keys.m_viewKeys = viewKeys;
             }
 
             if (!spendKeys && !viewKeys) {
-                this.m_spendKeys = new KeyPair();
-                this.m_viewKeys = new KeyPair(undefined, this.m_spendKeys.privateKey, undefined, 1);
+                keys.m_spendKeys = await KeyPair.from();
+                keys.m_viewKeys = await KeyPair.from(
+                    undefined, keys.m_spendKeys.privateKey, undefined, 1);
             }
+
+            return keys;
         }
 
         /**
          * Returns the spend keys
          */
-        public get spend(): KeyPair {
+        public get spend (): KeyPair {
             return this.m_spendKeys;
         }
 
@@ -142,14 +176,14 @@ export namespace ED25519 {
          * Sets the spend keys
          * @param keys
          */
-        public set spend(keys: KeyPair) {
+        public set spend (keys: KeyPair) {
             this.m_spendKeys = keys;
         }
 
         /**
          * Returns the view keys
          */
-        public get view(): KeyPair {
+        public get view (): KeyPair {
             return this.m_viewKeys;
         }
 
@@ -157,27 +191,24 @@ export namespace ED25519 {
          * Sets the view keys
          * @param keys
          */
-        public set view(keys: KeyPair) {
+        public set view (keys: KeyPair) {
             this.m_viewKeys = keys;
         }
     }
 }
 
 /** @ignore */
-function rand32(): string {
-    try {
-        return Mnemonics.random(256);
-    } catch (error) {
-        throw new Error('Could not retrieve 32-bytes of random data: ' + error.toString());
-    }
+function rand (bytes = 32): string {
+    return randomBytes(bytes)
+        .toString('base64');
 }
 
 /** @ignore */
-function simpleKdf(value: string, iterations: number): string {
+async function simpleKdf (value: string, iterations: number): Promise<string> {
     /** This is a very simple implementation of a pseudo PBKDF2 function */
-    let hex = Common.bin2hex(Common.str2bin(value));
+    let hex = Buffer.from(value).toString('hex');
     for (let i = 0; i < iterations; i++) {
-        hex = TurtleCoinCrypto.cn_fast_hash(hex);
+        hex = await TurtleCoinCrypto.cn_fast_hash(hex);
     }
     return hex;
 }
