@@ -2,10 +2,10 @@
 //
 // Please see the included LICENSE file for more information.
 
-import {Address} from './Address';
-import {ED25519} from './Types/ED25519';
-import {Interfaces, MultisigInterfaces, TransactionInputs, TurtleCoinCrypto} from './Types';
-import {Transaction} from './Transaction';
+import { Address } from './Address';
+import { ED25519 } from './Types/ED25519';
+import { Interfaces, MultisigInterfaces, TransactionInputs, TurtleCoinCrypto } from './Types';
+import { Transaction } from './Transaction';
 /** @ignore */
 import KeyPair = ED25519.KeyPair;
 
@@ -13,43 +13,42 @@ import KeyPair = ED25519.KeyPair;
  * Represents a multisig helper class that can be used for the creation of multisig wallets
  */
 export class Multisig {
-
     /**
      * Returns an address object representing the multisig wallet address
      */
-    public get address(): Address {
+    public async address (): Promise<Address> {
         if (!this.isReady) {
             throw new Error('Not all participants have been loaded');
         }
 
-        return Address.fromViewOnlyKeys(this.spend, this.view);
+        return Address.fromViewOnlyKeys(await this.spend(), await this.view());
     }
 
     /**
      * Returns the threshold (M) of the multisig wallet
      */
-    public get threshold(): number {
+    public get threshold (): number {
         return this.m_threshold;
     }
 
     /**
      * Returns the participants (N) of the multisig wallet
      */
-    public get participants(): number {
+    public get participants (): number {
         return this.m_participants;
     }
 
     /**
      * Returns the number of participants currently loaded into the object
      */
-    public get current_participants(): number {
+    public get current_participants (): number {
         return this.m_currentParticipants;
     }
 
     /**
      * Returns the shared private view key of the multisig wallet
      */
-    public get view(): string {
+    public async view (): Promise<string> {
         if (!this.isViewReady) {
             throw new Error('Not all participants have been loaded');
         }
@@ -60,20 +59,18 @@ export class Multisig {
     /**
      * Returns the shared public spend key of the multisig wallet
      */
-    public get spend(): string {
+    public async spend (): Promise<string> {
         if (!this.isSpendReady) {
             throw new Error('Not all participants have been loaded');
         }
 
         const keys: string[] = this.m_participant_keys;
 
-        this.m_multisig_keys.forEach((key) => {
-            if (key.publicKey.length !== 0 &&
-                keys.indexOf(key.publicKey) === -1
-            ) {
+        for (const key of this.m_multisig_keys) {
+            if (key.publicKey.length !== 0 && keys.indexOf(key.publicKey) === -1) {
                 keys.push(key.publicKey);
             }
-        });
+        }
 
         return TurtleCoinCrypto.calculateSharedPublicKey(keys);
     }
@@ -81,26 +78,28 @@ export class Multisig {
     /**
      * Returns if the object is ready for export and/or use
      */
-    public get isReady(): boolean {
+    public get isReady (): boolean {
         return (this.isViewReady && this.isSpendReady);
     }
 
     /**
      * Returns our multisig keys
      */
-    public get multisig_keys(): KeyPair[] {
+    public get multisig_keys (): KeyPair[] {
         return this.m_multisig_keys;
     }
 
     /**
      * Returns the public multisig keys
      */
-    public get public_multisig_keys(): string[] {
+    public async public_multisig_keys (): Promise<string[]> {
         const result: string[] = [];
 
-        this.calculated_multisig_keys.forEach((key) => {
+        const keys = await this.calculated_multisig_keys();
+
+        for (const key of keys) {
             result.push(key.publicKey);
-        });
+        }
 
         return result;
     }
@@ -108,12 +107,14 @@ export class Multisig {
     /**
      * Returns the private multisig keys
      */
-    public get private_multisig_keys(): string [] {
+    public async private_multisig_keys (): Promise<string []> {
         const result: string[] = [];
 
-        this.calculated_multisig_keys.forEach((key) => {
+        const keys = await this.calculated_multisig_keys();
+
+        for (const key of keys) {
             result.push(key.privateKey);
-        });
+        }
 
         return result;
     }
@@ -122,18 +123,19 @@ export class Multisig {
      * Calculates our multisig keys using the participant public spend keys
      * @returns our multisig keys key pairs
      */
-    private get calculated_multisig_keys(): KeyPair[] {
+    private async calculated_multisig_keys (): Promise<KeyPair[]> {
         if (this.m_threshold !== this.m_participants) {
             const multisig_keys: KeyPair[] = [];
 
-            this.m_wallet_multisig_keys.forEach((multisig_key) => {
+            for (const multisig_key of this.m_wallet_multisig_keys) {
                 const keys =
-                    TurtleCoinCrypto.calculateMultisigPrivateKeys(multisig_key.privateKey, this.m_participant_keys);
+                    await TurtleCoinCrypto.calculateMultisigPrivateKeys(
+                        multisig_key.privateKey, this.m_participant_keys);
 
-                keys.forEach((key) => {
-                    multisig_keys.push(new KeyPair(undefined, key));
-                });
-            });
+                for (const key of keys) {
+                    multisig_keys.push(await KeyPair.from(undefined, key));
+                }
+            }
 
             return multisig_keys;
         }
@@ -144,17 +146,17 @@ export class Multisig {
     /**
      * Returns if the view information is ready
      */
-    private get isViewReady(): boolean {
+    private get isViewReady (): boolean {
         return (this.m_currentParticipants === this.m_participants);
     }
 
     /**
      * Returns if the spend information is ready
      */
-    private get isSpendReady(): boolean {
-        const loaded = (this.threshold === this.participants) ?
-            this.m_participant_keys.length + 1 :
-            this.m_participant_keys.length;
+    private get isSpendReady (): boolean {
+        const loaded = (this.threshold === this.participants)
+            ? this.m_participant_keys.length + 1
+            : this.m_participant_keys.length;
         return (loaded === Multisig.requiredSigningKeys(this.threshold, this.participants));
     }
 
@@ -164,7 +166,7 @@ export class Multisig {
      * @param participants the wallet participants
      * @returns the number of additional exchange rounds required
      */
-    public static exchangeRoundsRequired(threshold: number, participants: number): number {
+    public static exchangeRoundsRequired (threshold: number, participants: number): number {
         return participants - threshold;
     }
 
@@ -175,7 +177,7 @@ export class Multisig {
      * @param participants the wallet participants
      * @returns a new instance of the object
      */
-    public static fromAddress(wallet: Address, threshold: number, participants: number): Multisig {
+    public static fromAddress (wallet: Address, threshold: number, participants: number): Multisig {
         if (!isValidThreshold(threshold, participants)) {
             throw new Error('Threshold does not require a majority of participants');
         }
@@ -211,29 +213,29 @@ export class Multisig {
      * @param participants the wallet participants
      * @returns a new instance of the object
      */
-    public static fromMultisigKeys(
+    public static async fromMultisigKeys (
         multisig_private_keys: string[],
         sharedPrivateViewKey: string,
         threshold: number,
-        participants: number,
-    ): Multisig {
+        participants: number
+    ): Promise<Multisig> {
         if (!isValidThreshold(threshold, participants)) {
             throw new Error('Threshold does not require a majority of participants');
         }
 
         const result = new Multisig();
 
-        multisig_private_keys.forEach((key) => {
-            if (!TurtleCoinCrypto.checkScalar(key)) {
+        for (const key of multisig_private_keys) {
+            if (!await TurtleCoinCrypto.checkScalar(key)) {
                 throw new Error('Found an invalid private key in the list of multisig private keys');
             }
 
-            result.m_multisig_keys.push(new KeyPair(undefined, key));
+            result.m_multisig_keys.push(await KeyPair.from(undefined, key));
 
-            result.m_wallet_multisig_keys.push(new KeyPair(undefined, key));
-        });
+            result.m_wallet_multisig_keys.push(await KeyPair.from(undefined, key));
+        }
 
-        if (!TurtleCoinCrypto.checkScalar(sharedPrivateViewKey)) {
+        if (!await TurtleCoinCrypto.checkScalar(sharedPrivateViewKey)) {
             throw new Error('Private view key is not a valid private key');
         }
 
@@ -252,7 +254,7 @@ export class Multisig {
      * @param participants the wallet participants
      * @returns the total number of signing keys created
      */
-    public static requiredSigningKeys(threshold: number, participants: number): number {
+    public static requiredSigningKeys (threshold: number, participants: number): number {
         return required_keys(threshold, participants);
     }
 
@@ -262,7 +264,7 @@ export class Multisig {
      * @param participants the wallet participants
      * @returns if the given scheme is valid for this library
      */
-    public static isValidThreshold(threshold: number, participants: number): boolean {
+    public static isValidThreshold (threshold: number, participants: number): boolean {
         return isValidThreshold(threshold, participants);
     }
 
@@ -274,11 +276,11 @@ export class Multisig {
      * @param partialKeyImages the partial key images
      * @returns the restored key image
      */
-    public static async restoreKeyImage(
+    public static async restoreKeyImage (
         publicEphemeral: string,
         derivation: string,
         outputIndex: number,
-        partialKeyImages: string[],
+        partialKeyImages: string[]
     ): Promise<string> {
         return TurtleCoinCrypto.restoreKeyImage(publicEphemeral, derivation, outputIndex, partialKeyImages);
     }
@@ -287,9 +289,9 @@ export class Multisig {
     private m_multisig_keys: KeyPair[] = [];
     private m_participant_keys: string[] = [];
     private m_view_keys: string[] = [];
-    private m_threshold: number = 0;
-    private m_participants: number = 0;
-    private m_currentParticipants: number = 1;
+    private m_threshold = 0;
+    private m_participants = 0;
+    private m_currentParticipants = 1;
 
     /**
      * Adds a participant to the multisig object
@@ -298,11 +300,11 @@ export class Multisig {
      * @param publicSpendKeys the participant spend key(s)
      * @param [privateViewKey] the private view key of the participant
      */
-    public async addParticipant(
+    public async addParticipant (
         publicSpendKeys: string[] | string,
-        privateViewKey?: string,
+        privateViewKey?: string
     ): Promise<void> {
-        if (privateViewKey && !TurtleCoinCrypto.checkScalar(privateViewKey)) {
+        if (privateViewKey && !await TurtleCoinCrypto.checkScalar(privateViewKey)) {
             throw new Error('Private view key is not a valid private key');
         }
 
@@ -314,21 +316,21 @@ export class Multisig {
             publicSpendKeys = [publicSpendKeys];
         }
 
-        publicSpendKeys.forEach((key) => {
-            if (!TurtleCoinCrypto.checkKey(key)) {
+        for (const key of publicSpendKeys) {
+            if (!await TurtleCoinCrypto.checkKey(key)) {
                 throw new Error('Found an invalid public spend key in the list');
             }
-        });
+        }
 
         if (privateViewKey && this.m_view_keys.indexOf(privateViewKey) === -1) {
             this.m_view_keys.push(privateViewKey);
         }
 
-        publicSpendKeys.forEach((key) => {
+        for (const key of publicSpendKeys) {
             if (this.m_participant_keys.indexOf(key) === -1) {
                 this.m_participant_keys.push(key);
             }
-        });
+        }
 
         this.m_currentParticipants++;
     }
@@ -340,10 +342,10 @@ export class Multisig {
      * @param outputIndex the index of the output in the transaction
      * @returns the partial key images
      */
-    public async generatePartialKeyImages(
+    public async generatePartialKeyImages (
         transactionHash: string,
         publicEphemeral: string,
-        outputIndex: number,
+        outputIndex: number
     ): Promise<MultisigInterfaces.PartialKeyImage[]> {
         const promises = [];
 
@@ -359,7 +361,7 @@ export class Multisig {
             partialKeyImages.push({
                 transactionHash,
                 outputIndex,
-                partialKeyImage: result,
+                partialKeyImage: result
             });
         }
 
@@ -371,8 +373,8 @@ export class Multisig {
      * @param tx the prepared transaction
      * @returns the partial signing keys
      */
-    public async generatePartialSigningKeys(
-        tx: Interfaces.PreparedTransaction,
+    public async generatePartialSigningKeys (
+        tx: Interfaces.PreparedTransaction
     ): Promise<MultisigInterfaces.PartialSigningKey[]> {
         const promises = [];
 
@@ -392,9 +394,9 @@ export class Multisig {
 
         for (const result of results) {
             partialSigningKeys.push({
-                transactionPrefixHash: tx.transaction.prefixHash,
+                transactionPrefixHash: await tx.transaction.prefixHash(),
                 index: result.index,
-                partialSigningKey: result.key,
+                partialSigningKey: result.key
             });
         }
 
@@ -407,9 +409,9 @@ export class Multisig {
      * @param partialSigningKeys the partial signing keys required for the signature scheme
      * @returns the completed transaction
      */
-    public async completeTransaction(
+    public async completeTransaction (
         tx: Interfaces.PreparedTransaction,
-        partialSigningKeys: MultisigInterfaces.PartialSigningKey[],
+        partialSigningKeys: MultisigInterfaces.PartialSigningKey[]
     ): Promise<Transaction> {
         const promises = [];
 
@@ -417,17 +419,15 @@ export class Multisig {
             const preparedRingSignature = getPreparedRingSignature(tx.signatureMeta, i);
             const ringPartialKeys = getPartialSigningKeys(partialSigningKeys, i);
 
-            for (const partialKey of ringPartialKeys) {
-                promises.push(restoreRingSignatures(
-                    preparedRingSignature.input.derivation,
-                    preparedRingSignature.input.outputIndex,
-                    ringPartialKeys,
-                    preparedRingSignature.realOutputIndex,
-                    preparedRingSignature.key,
-                    tx.transaction.signatures[i],
-                    i,
-                ));
-            }
+            promises.push(restoreRingSignatures(
+                preparedRingSignature.input.derivation,
+                preparedRingSignature.input.outputIndex,
+                ringPartialKeys,
+                preparedRingSignature.realOutputIndex,
+                preparedRingSignature.key,
+                tx.transaction.signatures[i],
+                i
+            ));
         }
 
         const results = await Promise.all(promises);
@@ -436,7 +436,7 @@ export class Multisig {
             tx.transaction.signatures[result.index] = result.sigs;
         }
 
-        const prefixHash = tx.transaction.prefixHash;
+        const prefixHash = await tx.transaction.prefixHash();
 
         const checkPromises = [];
 
@@ -445,7 +445,7 @@ export class Multisig {
                 prefixHash,
                 (tx.transaction.inputs[i] as TransactionInputs.KeyInput).keyImage,
                 getInputKeys(tx.signatureMeta, i),
-                tx.transaction.signatures[i],
+                tx.transaction.signatures[i]
             ));
         }
 
@@ -462,14 +462,14 @@ export class Multisig {
 }
 
 /** @ignore */
-function isValidThreshold(threshold: number, participants: number): boolean {
-    return (!((threshold / participants) <= .5) && required_keys(threshold, participants) <= 2 ** 13);
+function isValidThreshold (threshold: number, participants: number): boolean {
+    return (!((threshold / participants) <= 0.5) && required_keys(threshold, participants) <= 2 ** 13);
 }
 
 /** @ignore */
-function required_keys(
+function required_keys (
     threshold: number,
-    participants: number,
+    participants: number
 ): number {
     let result = participants;
 
@@ -485,23 +485,23 @@ function required_keys(
 }
 
 /** @ignore */
-async function generatePartialSigningKey(
+async function generatePartialSigningKey (
     preparedSignature: string,
     index: number,
-    privateSpendKey: string,
+    privateSpendKey: string
 ): Promise<{ key: string, index: number }> {
     const key = await TurtleCoinCrypto.generatePartialSigningKey(preparedSignature, privateSpendKey);
 
     return {
         key,
-        index,
+        index
     };
 }
 
 /** @ignore */
-function getRealOutputIndex(
+function getRealOutputIndex (
     tx: Interfaces.PreparedTransaction,
-    index: number,
+    index: number
 ): number {
     for (const sigs of tx.signatureMeta) {
         if (sigs.index === index) {
@@ -513,9 +513,9 @@ function getRealOutputIndex(
 }
 
 /** @ignore */
-function getPartialSigningKeys(
+function getPartialSigningKeys (
     partialSigningKeys: MultisigInterfaces.PartialSigningKey[],
-    index: number,
+    index: number
 ): MultisigInterfaces.PartialSigningKey[] {
     const results: MultisigInterfaces.PartialSigningKey[] = [];
 
@@ -529,9 +529,9 @@ function getPartialSigningKeys(
 }
 
 /** @ignore */
-function getPreparedRingSignature(
+function getPreparedRingSignature (
     preparedRingSignatures: Interfaces.PreparedRingSignature[],
-    index: number,
+    index: number
 ): Interfaces.PreparedRingSignature {
     for (const preparedRingSignature of preparedRingSignatures) {
         if (preparedRingSignature.index === index) {
@@ -543,14 +543,14 @@ function getPreparedRingSignature(
 }
 
 /** @ignore */
-async function restoreRingSignatures(
+async function restoreRingSignatures (
     derivation: string,
     outputIndex: number,
     partialSigningKeys: MultisigInterfaces.PartialSigningKey[],
     realOutputIndex: number,
     key: string,
     signatures: string[],
-    index: number,
+    index: number
 ): Promise<{ sigs: string[], index: number }> {
     const keys: string[] = [];
 
@@ -567,17 +567,17 @@ async function restoreRingSignatures(
         keys,
         realOutputIndex,
         key,
-        signatures,
+        signatures
     );
 
     return {
         sigs,
-        index,
+        index
     };
 }
 
 /** @ignore */
-function getInputKeys(preparedSignatures: Interfaces.PreparedRingSignature[], index: number): string[] {
+function getInputKeys (preparedSignatures: Interfaces.PreparedRingSignature[], index: number): string[] {
     for (const meta of preparedSignatures) {
         if (meta.index === index) {
             if (meta.inputKeys) {
@@ -590,11 +590,11 @@ function getInputKeys(preparedSignatures: Interfaces.PreparedRingSignature[], in
 }
 
 /** @ignore */
-async function checkRingSignatures(
+async function checkRingSignatures (
     hash: string,
     keyImage: string,
     publicKeys: string[],
-    signatures: string[],
+    signatures: string[]
 ): Promise<boolean> {
     return TurtleCoinCrypto.checkRingSignatures(hash, keyImage, publicKeys, signatures);
 }
