@@ -2,9 +2,10 @@
 //
 // Please see the included LICENSE file for more information.
 
-import {Reader, Writer} from 'bytestream-helper';
-import {ExtraNonceTag, TurtleCoinCrypto} from '../Types';
-import {Common} from '../Common';
+import { Reader, Writer } from 'bytestream-helper';
+import { ExtraNonceTag } from '../Types';
+import { Common } from '../Common';
+import { BigInteger } from 'big-integer';
 
 /** @ignore */
 enum SIZES {
@@ -23,6 +24,11 @@ export namespace ExtraTag {
         PUBKEY,
         NONCE,
         MERGED_MINING,
+        POW_NONCE,
+        RECIPIENT_PUBLIC_VIEW_KEY,
+        RECIPIENT_PUBLIC_SPEND_KEY,
+        TRANSACTION_PRIVATE_KEY,
+        POOL_NONCE
     }
 
     /**
@@ -52,22 +58,75 @@ export namespace ExtraTag {
         public abstract toString(): string;
     }
 
+    export class ExtraPowNonce implements IExtraTag {
+        public get tag (): ExtraTagType {
+            return this.m_tag;
+        }
+
+        public get size (): number {
+            return 9;
+        }
+
+        public static from (data: Buffer | string): ExtraPowNonce {
+            const reader = new Reader(data);
+
+            if (reader.varint().toJSNumber() !== ExtraTagType.POW_NONCE) {
+                throw new Error('Not a pow nonce field');
+            }
+
+            if (reader.unreadBytes !== SIZES.KEY) {
+                throw new RangeError('Not enough data available for reading');
+            }
+
+            const nonce = reader.uint64_t();
+
+            return new ExtraPowNonce(nonce);
+        }
+
+        private readonly m_tag: ExtraTagType = ExtraTagType.POW_NONCE;
+        private m_nonce: BigInteger;
+
+        constructor (nonce: BigInteger) {
+            this.m_nonce = nonce;
+        }
+
+        /**
+         * Represents the field as a Buffer
+         * @returns the Buffer representation of the object
+         */
+        public toBuffer (): Buffer {
+            const writer = new Writer();
+
+            writer.varint(this.tag);
+            writer.uint64_t(this.m_nonce);
+
+            return writer.buffer;
+        }
+
+        /**
+         * Represents the field as a hexadecimal string (blob)
+         * @returns the hexadecimal (blob) representation of the object
+         */
+        public toString (): string {
+            return this.toBuffer().toString('hex');
+        }
+    }
+
     /**
      * Represents a structured padding field used in the transaction extra field
      */
     export class ExtraPadding implements IExtraTag {
-
         /**
          * The tag type of the field
          */
-        public get tag(): ExtraTagType {
+        public get tag (): ExtraTagType {
             return this.m_tag;
         }
 
         /**
          * The size of the field in bytes including the tag
          */
-        public get size(): number {
+        public get size (): number {
             return this.toBuffer().length;
         }
 
@@ -77,7 +136,7 @@ export namespace ExtraTag {
          * @param data the data that makes up the nonce field
          * @returns the new object
          */
-        public static from(data: Buffer | string): ExtraPadding {
+        public static from (data: Buffer | string): ExtraPadding {
             const reader = new Reader(data);
 
             if (reader.varint().toJSNumber() !== ExtraTagType.PADDING) {
@@ -87,14 +146,14 @@ export namespace ExtraTag {
             return new ExtraPadding(0);
         }
 
-        public m_size: number = 0;
+        public m_size = 0;
         private readonly m_tag: ExtraTagType = ExtraTagType.PADDING;
 
         /**
          * Constructs a new field of the specified size
          * @param size the size of the padding field
          */
-        constructor(size: number) {
+        constructor (size: number) {
             this.m_size = size;
         }
 
@@ -102,7 +161,7 @@ export namespace ExtraTag {
          * Represents the field as a Buffer
          * @returns the Buffer representation of the object
          */
-        public toBuffer(): Buffer {
+        public toBuffer (): Buffer {
             const writer = new Writer();
 
             writer.varint(this.tag);
@@ -118,7 +177,7 @@ export namespace ExtraTag {
          * Represents the field as a hexadecimal string (blob)
          * @returns the hexadecimal (blob) representation of the object
          */
-        public toString(): string {
+        public toString (): string {
             return this.toBuffer().toString('hex');
         }
     }
@@ -127,25 +186,24 @@ export namespace ExtraTag {
      * Represents the transaction public key contained in the transaction extra field
      */
     export class ExtraPublicKey implements IExtraTag {
-
         /**
          * The tag type of the field
          */
-        public get tag(): ExtraTagType {
+        public get tag (): ExtraTagType {
             return this.m_tag;
         }
 
         /**
          * The public key contained in the field
          */
-        public get publicKey(): string {
+        public get publicKey (): string {
             return this.m_publicKey;
         }
 
         /**
          * The size of the field in bytes including the tag
          */
-        public get size(): number {
+        public get size (): number {
             return this.toBuffer().length;
         }
 
@@ -155,7 +213,7 @@ export namespace ExtraTag {
          * @param data the data that makes up the nonce field
          * @returns the new object
          */
-        public static from(data: Buffer | string): ExtraPublicKey {
+        public static from (data: Buffer | string): ExtraPublicKey {
             const reader = new Reader(data);
 
             if (reader.varint().toJSNumber() !== ExtraTagType.PUBKEY) {
@@ -178,19 +236,15 @@ export namespace ExtraTag {
          * Creates a new instance of the field using the supplied public key
          * @param publicKey the public key to be stored in the field
          */
-        constructor(publicKey: string) {
-            if (TurtleCoinCrypto.checkKey(publicKey)) {
-                this.m_publicKey = publicKey;
-            } else {
-                throw new Error('invalid publicKey supplied');
-            }
+        constructor (publicKey: string) {
+            this.m_publicKey = publicKey;
         }
 
         /**
          * Represents the field as a Buffer
          * @returns the Buffer representation of the object
          */
-        public toBuffer(): Buffer {
+        public toBuffer (): Buffer {
             const writer = new Writer();
 
             writer.varint(this.tag);
@@ -204,7 +258,7 @@ export namespace ExtraTag {
          * Represents the field as a hexadecimal string (blob)
          * @returns the hexadecimal (blob) representation of the object
          */
-        public toString(): string {
+        public toString (): string {
             return this.toBuffer().toString('hex');
         }
     }
@@ -213,32 +267,31 @@ export namespace ExtraTag {
      * Represents merged mining information contained in the transaction extra field
      */
     export class ExtraMergedMining implements IExtraTag {
-
         /**
          * The tag type of the field
          */
-        public get tag(): ExtraTagType {
+        public get tag (): ExtraTagType {
             return this.m_tag;
         }
 
         /**
          * The size of the field in bytes including the tag
          */
-        public get size(): number {
+        public get size (): number {
             return this.toBuffer().length;
         }
 
         /**
          * The depth of the block in the merkle root
          */
-        public get depth(): number {
+        public get depth (): number {
             return this.m_depth;
         }
 
         /**
          * The merkle root of the block merge mined
          */
-        public get merkleRoot(): string {
+        public get merkleRoot (): string {
             return this.m_merkleRoot;
         }
 
@@ -248,7 +301,7 @@ export namespace ExtraTag {
          * @param data the data that makes up the nonce field
          * @returns the new object
          */
-        public static from(data: Buffer | string): ExtraMergedMining {
+        public static from (data: Buffer | string): ExtraMergedMining {
             const reader = new Reader(data);
 
             if (reader.varint().toJSNumber() !== ExtraTagType.MERGED_MINING) {
@@ -293,7 +346,7 @@ export namespace ExtraTag {
          * @param depth The depth of the block in the merkle root
          * @param merkleRoot The merkle root of the block merge mined
          */
-        constructor(depth: number, merkleRoot: string) {
+        constructor (depth: number, merkleRoot: string) {
             this.m_depth = depth;
 
             if (!Common.isHex64(merkleRoot)) {
@@ -307,7 +360,7 @@ export namespace ExtraTag {
          * Represents the field as a Buffer
          * @returns the Buffer representation of the object
          */
-        public toBuffer(): Buffer {
+        public toBuffer (): Buffer {
             const writer = new Writer();
 
             writer.varint(this.tag);
@@ -329,7 +382,7 @@ export namespace ExtraTag {
          * Represents the field as a hexadecimal string (blob)
          * @returns the hexadecimal (blob) representation of the object
          */
-        public toString(): string {
+        public toString (): string {
             return this.toBuffer().toString('hex');
         }
     }
@@ -338,25 +391,24 @@ export namespace ExtraTag {
      * Represents nonce information contained in the transaction extra field
      */
     export class ExtraNonce implements IExtraTag {
-
         /**
          * The tag type of the field
          */
-        public get tag(): ExtraTagType {
+        public get tag (): ExtraTagType {
             return this.m_tag;
         }
 
         /**
          * The size of the field in bytes including the tag
          */
-        public get size(): number {
+        public get size (): number {
             return this.toBuffer().length;
         }
 
         /**
          * The Extra Nonce fields contained in the field
          */
-        public get tags(): IExtraNonce[] {
+        public get tags (): IExtraNonce[] {
             return this.m_tags;
         }
 
@@ -366,11 +418,11 @@ export namespace ExtraTag {
          * @param data the data that makes up the nonce field
          * @returns the new object
          */
-        public static from(data: Buffer | string): ExtraNonce {
+        public static from (data: Buffer | string): ExtraNonce {
             const reader = new Reader(data);
             const seen = {
                 paymentId: false,
-                data: false,
+                data: false
             };
 
             if (reader.varint().toJSNumber() !== ExtraTagType.NONCE) {
@@ -392,7 +444,7 @@ export namespace ExtraTag {
             const tags: ExtraNonceTag.IExtraNonce[] = [];
 
             while (reader.unreadBytes > 0) {
-                let tag: number = 0;
+                let tag = 0;
 
                 try {
                     tag = reader.varint(true).toJSNumber();
@@ -407,8 +459,12 @@ export namespace ExtraTag {
                     case ExtraNonceTag.NonceTagType.PAYMENT_ID:
                         totalLength += SIZES.KEY;
                         if (!seen.paymentId && reader.unreadBytes >= totalLength) {
-                            tags.push(ExtraNonceTag.ExtraNoncePaymentId.from(reader.bytes(33)));
-                            seen.paymentId = true;
+                            try {
+                                tags.push(ExtraNonceTag.ExtraNoncePaymentId.from(reader.bytes(33)));
+                                seen.paymentId = true;
+                            } catch (e) {
+                                reader.skip();
+                            }
                         } else {
                             reader.skip();
                         }
@@ -419,6 +475,10 @@ export namespace ExtraTag {
 
                             try {
                                 dataLength = reader.varint(true).toJSNumber();
+                                if (dataLength > reader.unreadBytes) {
+                                    reader.skip();
+                                    continue;
+                                }
                             } catch {
                                 reader.skip();
                                 continue;
@@ -427,8 +487,14 @@ export namespace ExtraTag {
                             totalLength += Common.varintLength(dataLength) + dataLength;
 
                             if (reader.unreadBytes >= totalLength) {
-                                tags.push(ExtraNonceTag.ExtraNonceData.from(reader.bytes(totalLength)));
-                                seen.data = true;
+                                try {
+                                    tags.push(ExtraNonceTag.ExtraNonceData.from(reader.bytes(totalLength)));
+                                    seen.data = true;
+                                } catch (e) {
+                                    reader.skip();
+                                }
+                            } else {
+                                reader.skip();
                             }
                         } else {
                             reader.skip();
@@ -450,7 +516,7 @@ export namespace ExtraTag {
          * Creates a new instance of the field using the supplied Extra Nonce fields
          * @param tags The Extra Nonce fields to place into the field
          */
-        constructor(tags: ExtraNonceTag.IExtraNonce[]) {
+        constructor (tags: ExtraNonceTag.IExtraNonce[]) {
             this.m_tags = tags;
         }
 
@@ -458,7 +524,7 @@ export namespace ExtraTag {
          * Remove a specific Extra Nonce field from the field
          * @param removeTag the Extra Nonce tag type to remove
          */
-        public removeTag(removeTag: ExtraNonceTag.NonceTagType) {
+        public removeTag (removeTag: ExtraNonceTag.NonceTagType) {
             const result: ExtraNonceTag.IExtraNonce[] = [];
 
             for (const tag of this.m_tags) {
@@ -476,7 +542,7 @@ export namespace ExtraTag {
          * Add/Update a specific Extra Nonce field to the field
          * @param tag the Extra Nonce tag to add/update
          */
-        public addTag(tag: ExtraNonceTag.IExtraNonce) {
+        public addTag (tag: ExtraNonceTag.IExtraNonce) {
             this.removeTag(tag.tag);
 
             this.tags.push(tag);
@@ -488,10 +554,20 @@ export namespace ExtraTag {
          * Represents the field as a Buffer
          * @returns the Buffer representation of the object
          */
-        public toBuffer(): Buffer {
+        public toBuffer (): Buffer {
             const writer = new Writer();
 
             writer.varint(this.tag);
+
+            const subWriter = new Writer();
+
+            for (const tag of this.tags) {
+                subWriter.write(tag.toBuffer());
+            }
+
+            writer.varint(subWriter.length);
+
+            writer.write(subWriter.blob);
 
             return writer.buffer;
         }
@@ -500,7 +576,343 @@ export namespace ExtraTag {
          * Represents the field as a hexadecimal string (blob)
          * @returns the hexadecimal (blob) representation of the object
          */
-        public toString(): string {
+        public toString (): string {
+            return this.toBuffer().toString('hex');
+        }
+    }
+
+    /**
+     * Represents the transaction private key contained in the transaction extra field
+     */
+    export class ExtraTransactionPrivateKey implements IExtraTag {
+        /**
+         * The tag type of the field
+         */
+        public get tag (): ExtraTagType {
+            return this.m_tag;
+        }
+
+        /**
+         * The public key contained in the field
+         */
+        public get privateKey (): string {
+            return this.m_privateKey;
+        }
+
+        /**
+         * The size of the field in bytes including the tag
+         */
+        public get size (): number {
+            return this.toBuffer().length;
+        }
+
+        /**
+         * Creates a new instance of the field using a Buffer or Blob copy of
+         * field created through other means
+         * @param data the data that makes up the nonce field
+         * @returns the new object
+         */
+        public static from (data: Buffer | string): ExtraTransactionPrivateKey {
+            const reader = new Reader(data);
+
+            if (reader.varint().toJSNumber() !== ExtraTagType.TRANSACTION_PRIVATE_KEY) {
+                throw new Error('Not a public key field');
+            }
+
+            if (reader.unreadBytes !== SIZES.KEY) {
+                throw new RangeError('Not enough data available for reading');
+            }
+
+            const privateKey = reader.hash();
+
+            return new ExtraTransactionPrivateKey(privateKey);
+        }
+
+        private readonly m_tag: ExtraTagType = ExtraTagType.TRANSACTION_PRIVATE_KEY;
+        private readonly m_privateKey: string = '';
+
+        /**
+         * Creates a new instance of the field using the supplied public key
+         * @param privateKey the public key to be stored in the field
+         */
+        constructor (privateKey: string) {
+            this.m_privateKey = privateKey;
+        }
+
+        /**
+         * Represents the field as a Buffer
+         * @returns the Buffer representation of the object
+         */
+        public toBuffer (): Buffer {
+            const writer = new Writer();
+
+            writer.varint(this.tag);
+
+            writer.hash(this.privateKey);
+
+            return writer.buffer;
+        }
+
+        /**
+         * Represents the field as a hexadecimal string (blob)
+         * @returns the hexadecimal (blob) representation of the object
+         */
+        public toString (): string {
+            return this.toBuffer().toString('hex');
+        }
+    }
+
+    /**
+     * Represents the recipient public view key contained in the transaction
+     * extra field for coinbase transactions
+     */
+    export class ExtraRecipientPublicViewKey implements IExtraTag {
+        /**
+         * The tag type of the field
+         */
+        public get tag (): ExtraTagType {
+            return this.m_tag;
+        }
+
+        /**
+         * The public key contained in the field
+         */
+        public get publicKey (): string {
+            return this.m_publicKey;
+        }
+
+        /**
+         * The size of the field in bytes including the tag
+         */
+        public get size (): number {
+            return this.toBuffer().length;
+        }
+
+        /**
+         * Creates a new instance of the field using a Buffer or Blob copy of
+         * field created through other means
+         * @param data the data that makes up the nonce field
+         * @returns the new object
+         */
+        public static from (data: Buffer | string): ExtraRecipientPublicViewKey {
+            const reader = new Reader(data);
+
+            if (reader.varint().toJSNumber() !== ExtraTagType.RECIPIENT_PUBLIC_VIEW_KEY) {
+                throw new Error('Not a public key field');
+            }
+
+            if (reader.unreadBytes !== SIZES.KEY) {
+                throw new RangeError('Not enough data available for reading');
+            }
+
+            const publicKey = reader.hash();
+
+            return new ExtraRecipientPublicViewKey(publicKey);
+        }
+
+        private readonly m_tag: ExtraTagType = ExtraTagType.RECIPIENT_PUBLIC_VIEW_KEY;
+        private readonly m_publicKey: string = '';
+
+        /**
+         * Creates a new instance of the field using the supplied public key
+         * @param publicKey the public key to be stored in the field
+         */
+        constructor (publicKey: string) {
+            this.m_publicKey = publicKey;
+        }
+
+        /**
+         * Represents the field as a Buffer
+         * @returns the Buffer representation of the object
+         */
+        public toBuffer (): Buffer {
+            const writer = new Writer();
+
+            writer.varint(this.tag);
+
+            writer.hash(this.publicKey);
+
+            return writer.buffer;
+        }
+
+        /**
+         * Represents the field as a hexadecimal string (blob)
+         * @returns the hexadecimal (blob) representation of the object
+         */
+        public toString (): string {
+            return this.toBuffer().toString('hex');
+        }
+    }
+
+    /**
+     * Represents the recipient public spend key contained in the transaction
+     * extra field for coinbase transactions
+     */
+    export class ExtraRecipientPublicSpendKey implements IExtraTag {
+        /**
+         * The tag type of the field
+         */
+        public get tag (): ExtraTagType {
+            return this.m_tag;
+        }
+
+        /**
+         * The public key contained in the field
+         */
+        public get publicKey (): string {
+            return this.m_publicKey;
+        }
+
+        /**
+         * The size of the field in bytes including the tag
+         */
+        public get size (): number {
+            return this.toBuffer().length;
+        }
+
+        /**
+         * Creates a new instance of the field using a Buffer or Blob copy of
+         * field created through other means
+         * @param data the data that makes up the nonce field
+         * @returns the new object
+         */
+        public static from (data: Buffer | string): ExtraRecipientPublicSpendKey {
+            const reader = new Reader(data);
+
+            if (reader.varint().toJSNumber() !== ExtraTagType.RECIPIENT_PUBLIC_SPEND_KEY) {
+                throw new Error('Not a public key field');
+            }
+
+            if (reader.unreadBytes !== SIZES.KEY) {
+                throw new RangeError('Not enough data available for reading');
+            }
+
+            const publicKey = reader.hash();
+
+            return new ExtraRecipientPublicSpendKey(publicKey);
+        }
+
+        private readonly m_tag: ExtraTagType = ExtraTagType.RECIPIENT_PUBLIC_SPEND_KEY;
+        private readonly m_publicKey: string = '';
+
+        /**
+         * Creates a new instance of the field using the supplied public key
+         * @param publicKey the public key to be stored in the field
+         */
+        constructor (publicKey: string) {
+            this.m_publicKey = publicKey;
+        }
+
+        /**
+         * Represents the field as a Buffer
+         * @returns the Buffer representation of the object
+         */
+        public toBuffer (): Buffer {
+            const writer = new Writer();
+
+            writer.varint(this.tag);
+
+            writer.hash(this.publicKey);
+
+            return writer.buffer;
+        }
+
+        /**
+         * Represents the field as a hexadecimal string (blob)
+         * @returns the hexadecimal (blob) representation of the object
+         */
+        public toString (): string {
+            return this.toBuffer().toString('hex');
+        }
+    }
+
+    /**
+     * Represents the extra field for the pool nonce used by mining pools
+     */
+    export class ExtraPoolNonce implements IExtraTag {
+        /**
+         * The tag type of the field
+         */
+        public get tag (): ExtraTagType {
+            return this.m_tag;
+        }
+
+        /**
+         * The size of the field in bytes including the tag
+         */
+        public get size (): number {
+            return this.toBuffer().length;
+        }
+
+        /**
+         * The arbitrary data included in the field
+         */
+        public get data (): Buffer {
+            return this.m_data;
+        }
+
+        /**
+         * Creates a new instance of the field using a Buffer or Blob copy of
+         * field created through other means
+         * @param data the data that makes up the nonce field
+         * @returns the new object
+         */
+        public static from (data: Buffer | string): ExtraPoolNonce {
+            const reader = new Reader(data);
+
+            if (reader.varint().toJSNumber() !== ExtraTagType.POOL_NONCE) {
+                throw new Error('Not an extra data field');
+            }
+
+            try {
+                reader.varint(true);
+            } catch {
+                throw new Error('Cannot read required field data');
+            }
+
+            const length = reader.varint().toJSNumber();
+
+            if (reader.unreadBytes !== length) {
+                throw new RangeError('Not enough data available for reading');
+            }
+
+            const dataLength = reader.bytes(length);
+
+            return new ExtraPoolNonce(dataLength);
+        }
+
+        private readonly m_tag: ExtraTagType = ExtraTagType.POOL_NONCE;
+        private readonly m_data: Buffer = Buffer.alloc(0);
+
+        /**
+         * Creates a new instance of the field from just the extra data to be included
+         * @param data the pool nonce data to be included
+         */
+        constructor (data: Buffer) {
+            this.m_data = data;
+        }
+
+        /**
+         * Represents the field as a Buffer
+         * @returns the Buffer representation of the object
+         */
+        public toBuffer (): Buffer {
+            const writer = new Writer();
+
+            writer.varint(this.tag);
+
+            writer.varint(this.data.length);
+
+            writer.write(this.data);
+
+            return writer.buffer;
+        }
+
+        /**
+         * Represents the field as a hexadecimal string (blob)
+         * @returns the hexadecimal (blob) representation of the object
+         */
+        public toString (): string {
             return this.toBuffer().toString('hex');
         }
     }
