@@ -2,8 +2,8 @@
 //
 // Please see the included LICENSE file for more information.
 
-import {Reader, Writer} from 'bytestream-helper';
-import {BigInteger, PortableStorageConstants} from '../Types';
+import { Reader, Writer } from 'bytestream-helper';
+import { BigInteger, PortableStorageConstants } from '../Types';
 
 /** @ignore */
 type PortableStoreValue =
@@ -45,16 +45,15 @@ interface PortableStorageEntry {
 
 /** @ignore */
 export class PortableStorage {
-
-    public get version(): number {
+    public get version (): number {
         return this.m_version;
     }
 
-    public set version(value: number) {
+    public set version (value: number) {
         this.m_version = value;
     }
 
-    public static from(data: Reader | Buffer | string, skipHeader: boolean = false): PortableStorage {
+    public static from (data: Reader | Buffer | string, skipHeader = false): PortableStorage {
         const reader = new Reader(data);
 
         const result = new PortableStorage();
@@ -67,8 +66,8 @@ export class PortableStorage {
             result.m_signatureA = reader.uint32_t().toJSNumber();
             result.m_signatureB = reader.uint32_t().toJSNumber();
 
-            if (result.m_signatureA !== PortableStorageConstants.SIGNATURE_A
-                || result.m_signatureB !== PortableStorageConstants.SIGNATURE_B) {
+            if (result.m_signatureA !== PortableStorageConstants.SIGNATURE_A ||
+                result.m_signatureB !== PortableStorageConstants.SIGNATURE_B) {
                 throw new Error('Portable storage signature failure');
             }
 
@@ -89,7 +88,7 @@ export class PortableStorage {
     protected m_signatureB: number = PortableStorageConstants.SIGNATURE_B;
     protected m_entries: PortableStorageEntry[] = [];
 
-    public get(name: string): PortableStoreValue {
+    public get (name: string): PortableStoreValue {
         for (const entry of this.m_entries) {
             if (entry.name === name) {
                 return entry.value;
@@ -99,7 +98,7 @@ export class PortableStorage {
         throw new Error('Entry not found');
     }
 
-    public set(name: string, value: PortableStoreValue, type: StorageType) {
+    public set (name: string, value: PortableStoreValue, type: StorageType) {
         // tslint:disable-next-line:prefer-for-of
         for (let i = 0; i < this.m_entries.length; i++) {
             if (this.m_entries[i].name === name && this.m_entries[i].type === type) {
@@ -108,10 +107,10 @@ export class PortableStorage {
             }
         }
 
-        this.m_entries.push({name, type, value});
+        this.m_entries.push({ name, type, value });
     }
 
-    public exists(name: string): boolean {
+    public exists (name: string): boolean {
         for (const entry of this.m_entries) {
             if (entry.name === name) {
                 return true;
@@ -121,7 +120,7 @@ export class PortableStorage {
         return false;
     }
 
-    public toBuffer(skipHeader: boolean = false): Buffer {
+    public toBuffer (skipHeader = false): Buffer {
         const writer = new Writer();
 
         if (!skipHeader) {
@@ -135,13 +134,13 @@ export class PortableStorage {
         return writer.buffer;
     }
 
-    public toString(skipHeader: boolean = false): string {
+    public toString (skipHeader = false): string {
         return this.toBuffer(skipHeader).toString('hex');
     }
 }
 
 /** @ignore */
-function blobToEntries(reader: Reader): PortableStorageEntry[] {
+function blobToEntries (reader: Reader): PortableStorageEntry[] {
     const entryCount = reader.varint(false, true).toJSNumber();
     const entries: PortableStorageEntry[] = [];
 
@@ -157,45 +156,45 @@ function blobToEntries(reader: Reader): PortableStorageEntry[] {
 
         switch (type) {
             case StorageType.NULL:
-                entries.push({name, type, value: null});
+                entries.push({ name, type, value: null });
                 break;
 
             case StorageType.UINT64:
-                entries.push({name, type, value: reader.uint64_t(true)});
+                entries.push({ name, type, value: reader.uint64_t(true) });
                 break;
 
             case StorageType.UINT32:
-                entries.push({name, type, value: reader.uint32_t()});
+                entries.push({ name, type, value: reader.uint32_t() });
                 break;
 
             case StorageType.UINT16:
-                entries.push({name, type, value: reader.uint16_t()});
+                entries.push({ name, type, value: reader.uint16_t() });
                 break;
 
             case StorageType.UINT8:
-                entries.push({name, type, value: reader.uint8_t()});
+                entries.push({ name, type, value: reader.uint8_t() });
                 break;
 
             case StorageType.STRING:
                 valueLength = reader.varint(false, true).toJSNumber();
-                entries.push({name, type, value: reader.bytes(valueLength).toString('hex')});
+                entries.push({ name, type, value: reader.bytes(valueLength).toString('hex') });
                 break;
 
             case StorageType.BOOL:
                 value = reader.uint8_t().toJSNumber();
-                entries.push({name, type, value: (value === 1)});
+                entries.push({ name, type, value: (value === 1) });
                 break;
 
             case StorageType.OBJECT:
-                entries.push({name, type, value: PortableStorage.from(reader, true)});
+                entries.push({ name, type, value: PortableStorage.from(reader, true) });
                 break;
 
             case StorageType.BUFFER:
                 valueLength = reader.varint(false, true).toJSNumber();
-                entries.push({name, type, value: reader.bytes(valueLength)});
+                entries.push({ name, type, value: reader.bytes(valueLength) });
                 break;
 
-            case StorageType.STRING_ARRAY:
+            case StorageType.STRING_ARRAY: {
                 arrCount = reader.varint(false, true).toJSNumber();
                 const sa_arr: string[] = [];
 
@@ -205,10 +204,11 @@ function blobToEntries(reader: Reader): PortableStorageEntry[] {
                     sa_arr.push(reader.bytes(stringLength).toString('hex'));
                 }
 
-                entries.push({name, type, value: sa_arr});
+                entries.push({ name, type, value: sa_arr });
+            }
                 break;
 
-            case StorageType.OBJECT_ARRAY:
+            case StorageType.OBJECT_ARRAY: {
                 arrCount = reader.varint(false, true).toJSNumber();
                 const oa_arr: PortableStorage[] = [];
 
@@ -216,7 +216,8 @@ function blobToEntries(reader: Reader): PortableStorageEntry[] {
                     oa_arr.push(PortableStorage.from(reader, true));
                 }
 
-                entries.push({name, type, value: oa_arr});
+                entries.push({ name, type, value: oa_arr });
+            }
                 break;
             default:
                 throw new Error(type + ' not implemented');
@@ -227,7 +228,7 @@ function blobToEntries(reader: Reader): PortableStorageEntry[] {
 }
 
 /** @ignore */
-function entriesToBuffer(entries: PortableStorageEntry[]): Buffer {
+function entriesToBuffer (entries: PortableStorageEntry[]): Buffer {
     const writer = new Writer();
 
     writer.varint(entries.length, true);
@@ -251,10 +252,11 @@ function entriesToBuffer(entries: PortableStorageEntry[]): Buffer {
             case StorageType.UINT8:
                 writer.uint8_t(entry.value);
                 break;
-            case StorageType.STRING:
+            case StorageType.STRING: {
                 const s_buf = Buffer.from(entry.value, 'hex');
                 writer.varint(s_buf.length, true);
                 writer.write(s_buf);
+            }
                 break;
             case StorageType.BOOL:
                 writer.uint8_t((entry.value) ? 1 : 0);
